+++ AprioritTest/TransactionTracer/src/client/BtcApiClient.h
@@ -11,6 +11,6 @@ using std::string;
 class BtcApiClient {
 	const string host = "https://blockchain.info/";
 public:
-	std::future<void> getTxRawAsync(int txid, std::function<void (cpr::Response)>&& cb);
+	std::future<void> getTxRawAsync(uint64_t txid, std::function<void (cpr::Response)>&& cb);
 	cpr::Response getTxRaw(string tx_hash);
 };
